brush: refactor splitting Body component into three new files
