+++ brush/src/pages/Main/EmptyRom.js
@@ -1,0 +1,6 @@
+import React from 'react'
+import InputRom from '../../components/InputRom'
+
+const EmptyRom = () => <InputRom />
+
+export default EmptyRom
