+++ brush/src/pages/Main/index.js
@@ -1,0 +1,18 @@
+import React, { useContext } from 'react'
+import ROMContext from '../../context/ROMContext'
+import EmptyRom from './EmptyRom'
+import LoadedRom from './LoadedRom'
+
+const Main = () => {
+  const { romBufferStatus } = useContext(ROMContext)
+
+  const contentStates = {
+    empty: EmptyRom,
+    loaded: LoadedRom,
+    starting: () => null,
+  }
+
+  return React.createElement(contentStates[romBufferStatus])
+}
+
+export default Main
